Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-definition/grid-change-fit-content-argument-001-expected.txt

CONSOLE MESSAGE: line 190: No valid data-* attributes found in selector list : .grid
CONSOLE MESSAGE: line 190: No valid data-* attributes found in selector list : .grid
CONSOLE MESSAGE: line 190: No valid data-* attributes found in selector list : .grid
XXXX XXX XX X X
XXXX XXX XX X X

PASS .grid 1 
PASS .grid 2 
